Move magic ISBN, RFC, and GEO links to between replaceExternalLinks and replaceInternalLinks. This is a basic fix for the infamous bug [ 708163 ] [[ISBN #]] doesn't work correctly
ISBN, RFC links accidentally put inside [[ ]] brackets will now come up
relatively straightforwardly, with extra brackets. This is much nicer
than a horror of HTML errors that there were before; does not solve
problem with embedded <a>-in-<a> tag if you put it in a link piped
text, but that's a smaller problem.

Someday a full tokenized parser will wish this problem away into the
cornfields.

svn path=/trunk/phase3/; revision=4506