In the allpages index, make the whole line links to hopefully reduce the confusion factor. Some people thought that the link was actually a link _to_ the first page listed.
Since the centering is done with a table (yecch) each line is
actually three separate links, for the first page, "to", and
last page. All go to the same place, and on Monobook with link
underlining off it should look pretty decent. It's less attractive
on the classic skins with underlining on, but it's better than
the old look i think.

Also, removed the old caching with LogPage, since this no longer
works. Either the index generation needs to be made a lot faster,
or the index points should be saved in the querycache table.

http://bugzilla.wikipedia.org/show_bug.cgi?id=250

svn path=/trunk/phase3/; revision=4978