ExternalEditor helper reference implementation (pre-alpha)
This helper application will fetch a resource such as a wiki page or an
image, load it into an application like a text editor or an image processor,
and save it back to the server. It can be associated with the MIME type
application/external-editor to be directly used by supporting software
like MediaWiki.

svn path=/trunk/phase3/; revision=7858