* (bug 3348) Some additional weak password checks: password which is same as username will now be rejected.
There are some UI issues still with handling of bad/weak passwords. It's not too clear to the legit user who tries to log in again. There's also some bad behavior on the create account form; short or bad passwords just give an exception backtrace. Nice. :) But that's existing problems...

svn path=/trunk/phase3/; revision=21640