* (bug 10971) `aufrom` parameter doesn't work with spaces * Fix username handling issue with `auprefix` parameter
Usernames are *not* stored in "key" format, and treating them as such will cause various comparisons to misbehave. This was preventing `aufrom` from including the username (as it was supposed to) and would have broken `auprefix` for all comparisons containing spaces.

As a side-effect, both these parameters will now treat underscores as equivalent to spaces.

svn path=/trunk/phase3/; revision=24870