* Don't die when single-element arrays are passed to SQL query constructors   that have an array index other than 0
Since we don't enforce linear ordering of array keys on longer arrays, we
may as well not enforce it on single-element arrays either.
The freaky little Categoryfinder class was using different keys for its
arrays, which could cause a PHP notice and silent lookup failure when
only one ID was looked up.


svn path=/trunk/phase3/; revision=32452