* (bug 12644) Template list on edit page now sorted on preview
On initial edit, items were sorted by virtue of being pulled from a sorted table, but on preview they were displayed in rough order of use in parsing, which is a bit less legible.

Added Title::compare() static comparator available as a callback for usort(), using it in Linker::formatTemplates().

svn path=/trunk/phase3/; revision=34485