Revert r39055 for now "Optional 'edit notices' on a per-namespace and per-page basis, per feature request 15102"
Sanitizer::escapeClass() isn't really suitable for constructing message key names; a lot of names won't be easy to predict and it would be difficult for people to set up a custom message for, say, a page name containing punctuation or non-ASCII characters.

The check for message availability is using wfMsgForContent() for the content language, but then the output is done via $wgOut->addWikiMsg() which will use the UI language. Is this the desired behavior?


svn path=/trunk/phase3/; revision=39081