Prohibit empty page titles at a low level
This adds a sanity check to EditPage::doEdit() that throws an exception if the Title's name (sans namespace) is empty.  Apparently the API edit module doesn't handle this error correctly at a high level, as evidenced by page 19405691 on enwiki.  I didn't try to test whether this extra check stops the particular error, but it doesn't hurt in any case.

svn path=/trunk/phase3/; revision=41085