* (bug 13342) importScript() generates more consistent URI encoding
Switches order of URL to put title first, and encodes ':' and '/' consistently with MediaWiki general URI encoding.


The current implementation escapes ':' and '/', producing URLs like this:

/trunk/index.php?action=raw&ctype=text/javascript&title=MediaWiki%3ASysop.js

If we're recommending consistent use of importScript() it probably doesn't really matter, but it wouldn't hurt to be self-consistent with how we generate other URLs... of course even with the change to the encoding, the order of the pieces is different.

MediaWiki usually generates URLs with the title component first; also caching options will differ depending on who's generating the URL and when.

svn path=/trunk/phase3/; revision=44192