(bug 16950) Show move log when recreating a page
Showing the move log as well as delete log helps with pages that were
moved with no redirect, or moved and then the redirect deleted with a
confusing message.  Previously, such pages would appear to have never
existed to a user trying to create them.

Patch by church of emacs, with trivial modifications by me.

svn path=/trunk/phase3/; revision=49821