Remove KHTMLFixes.css
This was added more than five years ago, in r3532.  It was being loaded
for WebKit as well as KHTML, and I could notice no difference in either
Chrome or Konqueror whether it was loaded or not.  It was responsible
for causing display errors with an HTML 5 doctype in WebKit.

svn path=/trunk/phase3/; revision=53141