Re-enable $wgHtml5
The rendering issue that prompted r53137 has been fixed in r53141.
However, everyone should be on the lookout for browsers becoming
slightly more standards-compliant when this is enabled, possibly
breaking some things.  Since there are no known issues, it should be
safe to re-enable so it can get further testing.  See this page for info
on the effects of different doctypes:

http://hsivonen.iki.fi/doctype/

svn path=/trunk/phase3/; revision=53142