More convenient format for $wgExternalAuthType
Now is the whole class name, not a suffix.  This allows you to do things
like $wgExternalAuthType::someStaticMethod() instead of having to fiddle
with temporary variables.

This will obviously break anything that has $wgExternalAuthType already
configured, so if that's anyone but me, make sure to fix your configs.
This is still experimental.

svn path=/trunk/phase3/; revision=60017