Do not add the license selector when uploading a new version of a file
When using "upload a new version", the information in the license field is ignored during submit.
The field has no use in this mode because the information is only added in the filelog, not on
the filepage. As such we should just not add it at all when uploading a new version.

This closes bug 23241.



svn path=/trunk/phase3/; revision=65899