new-installer: GRANT permissions to our new non-root user by sourcing users.sql
The old installer sourced maintenance/users.sql to create a new
non-root user if requested. Do that in the new installer as well.

So far this only works on MySQL, but I'm adding a generic setupUser
function on the assumption that other databases (except SQLite) want
to do some sort of user setup.

svn path=/trunk/phase3/; revision=69008