Revert r74122, making r74117 live again
Some refactoring, to encapsulate users. Added a UserWrapper, contains the setup username/password and user object. Allows for ease of adding of multiple users (and groups)

Per Brion, if logging in user == wguser, we skip login tests, so use a secondary user

Fixup testApiLoginBadPass, we should get a WrongPass when the passwords wrong, but we've given a token

svn path=/trunk/phase3/; revision=74213