js email validation fixed + i18n support
The Javascript email validation in User:Preference got broken by an id change,
this patch fix the issue.

A long standing issue was a lack of proper i18n support, message helper
was hard coded in the js file.  It now use the resourceloader to get the
i18nized message.

Follow up r75670


svn path=/trunk/phase3/; revision=79904